chore(protocol): move protocol runtime into packages/protocol
Moves serializers.ts, validator.ts and validatorPrimitives.ts out of
packages/playwright-core/src/protocol into packages/protocol/src,
next to the existing structs.d.ts. Consumers now import these via the
@protocol/* alias, and the generated validator.ts is emitted to the
new location.
