--- packages/playwright-core/src/protocol/DEPS.list
@@ -1,1 +1,0 @@
-[*]
