+++ packages/playwright-core/src/server/dispatchers/DEPS.list
@@ -1,7 +1,7 @@
 [*]
 ../../generated/
-../../protocol/
 @utils/**
 @isomorphic/**
+@protocol/**
 ../**
 node_modules/yaml
